spotdl/requirements.txt
django
djangorestframework
pydantic
python-telegram-bot